refactor: :fire: Remove save_load
Removing this since people will be clicking the save button instead. The packed_qty > pick_qty check should be added back in at some point. I tried to added it to the write function but it wasn't behaving exactly as intended when clicking the save button. So just decided to remove it for now

#89